+++ app/pokesearch/(overview)/loading.tsx
@@ -1,5 +1,5 @@
+import PokeSearchSkeleton from '@/app/ui/utils/skeletons'
 import { type ReactElement } from 'react'
-import PokeSearchSkeleton from '@/app/ui/skeletons'
 
 export default function Loading (): ReactElement {
   return <PokeSearchSkeleton/>
