+++ app/ui/header/navbar.tsx
@@ -3,7 +3,7 @@ import { type ReactElement, useState, useEffect } from 'react'
 import { Bars3Icon, HomeIcon, InformationCircleIcon, MagnifyingGlassIcon, XMarkIcon } from '@heroicons/react/24/outline'
 import Link from 'next/link'
 import clsx from 'clsx'
-import Logo from './logo'
+import Logo from '../utils/logo'
 import ThemeMenu from './theme-menu'
 import NavLinks, { type NavLinkContent } from './nav-links'
 import { useDebouncedCallback } from 'use-debounce'
